More lint cleanup
