+++ Sources/GenIR/Extensions/URL+Extensions.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ArgumentParser
 
 extension URL {
 	/// Returns the path component of a URL
@@ -15,5 +16,11 @@ extension URL {
 
 	var fileExists: Bool {
 		return FileManager().fileExists(atPath: self.filePath)
+	}
+}
+
+extension URL: ExpressibleByArgument {
+	public init?(argument: String) {
+		self = argument.fileURL.absoluteURL
 	}
 }
